## What changes are proposed in this pull request?
The ordering and pagination of the requested runs is done in the SQL backend

## How is this patch tested?
Unit Tests
performance test on a dump of a real prod DB: 
- requests.post('....../runs/search', json='{"experiment_ids":["1"]}') 
------ ~14.5 s -> ~5.8s
- requests.post('....../runs/search', json='{"experiment_ids":["1"], "filter":"metrics.triplet_precision >= 0.87"}') ~0.11 s -> ~0.11s
- requests.post('....../runs/search', json='{"experiment_ids":["1"], "max_results": 10, "filter":"attribute.status = \'running\' "}') ~1.32s -> ~0.075s

## Release Notes
performance improvement in the search_runs endpoint


### Is this a user-facing change?
perf only
- [ ] No. You can skip the rest of this section.
- [ ] Yes. Give a description of this change to be included in the release notes for MLflow users.

### What component(s) does this PR affect?

- [ ] UI
- [ ] CLI
- [ ] API
- [ ] REST-API
- [ ] Examples
- [ ] Docs
- [x] Tracking
- [ ] Projects
- [ ] Artifacts
- [ ] Models
- [ ] Scoring
- [ ] Serving
- [ ] R
- [ ] Java
- [ ] Python

### How should the PR be classified in the release notes? Choose one:

- [ ] `rn/breaking-change` - The PR will be mentioned in the "Breaking Changes" section
- [ ] `rn/none` - No description will be included. The PR will be mentioned only by the PR number in the "Small Bugfixes and Documentation Updates" section
- [ ] `rn/feature` - A new user-facing feature worth mentioning in the release notes
- [ ] `rn/bug-fix` - A user-facing bug fix worth mentioning in the release notes
- [ ] `rn/documentation` - A user-facing documentation change worth mentioning in the release notes