@jcuquemelle 2019-11-13
> This error doesn't seem to occur when the orderby clauses reference _distinct_ tables. Can you take a look and see what might be going on?

The issue was that the name conflict was in fact occuring between all CASE WHEN columns, using labels on each subquery would solve between-tables conflicts only. The new solution reverts labelling columns in the subquery but assigns a unique label for each CASE WHEN clause