@jcuquemelle 2019-11-14
close and reopen to relaunch Travis (got a iris_training.csv: 503 -- Service Unavailable error in py 2.7 tests)